refactor: split tree classes into different modules
